+++ apps/docs/src/app/core/api-files.ts
@@ -114,23 +114,6 @@ export const API_FILES = {
         'ListFooterDirective',
         'ListMessageDirective'
     ],
-    localizationEditor: [
-        'LocalizationEditorComponent',
-        'LocalizationEditorItemComponent',
-        'LocalizationEditorMainComponent',
-        'LocalizationEditorInputDirective',
-        'LocalizationEditorTextareaDirective',
-        'LocalizationEditorLabel'
-    ],
-    megaMenu: [
-        'MegaMenuComponent',
-        'MegaMenuGroupComponent',
-        'MegaMenuItemComponent',
-        'MegaMenuLinkDirective',
-        'MegaMenuTitleDirective',
-        'MegaMenuSubitemDirective',
-        'MegaMenuSublinkDirective'
-    ],
     menu: [
         'MenuAddonDirective',
         'MenuComponent',
@@ -299,7 +282,8 @@ export const API_FILES = {
         'ToolbarComponent',
         'ToolbarContentComponent',
         'ToolbarOverflowContentComponent',
-        'ToolbarSpacerComponent'
+        'ToolbarSpacerComponent',
+        'ThemesService'
     ],
     tree: ['TreeComponent', 'TreeChildComponent', 'TreeRowObjectModel'],
     wizard: ['WizardComponent']