+++ libs/core/src/lib/feed-list-item/feed-list-item.module.ts
@@ -1,20 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { PipeModule } from '../utils/public_api';
+import { PipeModule, LineClampModule } from '../utils/public_api';
 import { LinkModule } from '../link/public_api';
 import { FormattedTextModule } from '../formatted-text/public_api';
 import { FeedListItemComponent } from './components/item/feed-list-item.component';
 import { FeedListComponent } from './components/list/feed-list.component';
-import { LineClampDirective, LineClampTargetDirective } from './directives/line-clamp.directive';
 import { FeedListAvatarComponent } from './components/feed-list-avatar/feed-list-avatar.component';
 import { FeedListActionComponent } from './components/feed-list-action/feed-list-action.component';
 import { FeedListFooterComponent } from './components/feed-list-footer/feed-list-footer.component';
 
 @NgModule({
   declarations: [
-    LineClampDirective,
-    LineClampTargetDirective,
     FeedListComponent,
     FeedListItemComponent,
     FeedListAvatarComponent,
@@ -25,7 +22,8 @@ import { FeedListFooterComponent } from './components/feed-list-footer/feed-list
     CommonModule,
     PipeModule,
     FormattedTextModule,
-    LinkModule
+    LinkModule,
+    LineClampModule
   ],
   exports: [
     FeedListComponent,